mambanegra/examples/requirements.locked.txt
cassandra-driver==3.7.1
futures==3.0.5
lru-dict==1.1.6
pymongo==3.4.0
six==1.10.0
